requirements.txt
behave
gherkin3
allure-behave
openpyxl
jira
mysql
mysql-connector-python
selenium
webdriver-manager
pandas
requests
cx_Oracle
XlsxWriter
PyHamcrest
pymongo
cassandra-driver
gql
sqlalchemy
Appium-Python-Client
elasticsearch
boto3
kubernetes